Bump SDK version. Update TS wrappers
