+++ RNZumoKit.podspec
@@ -2,7 +2,7 @@
 
 Pod::Spec.new do |s|
   s.name         = "RNZumoKit"
-  s.version      = "4.0.0-beta.10"
+  s.version      = "4.1.0-alpha.2"
   s.summary      = "RNZumoKit"
   s.description  = <<-DESC
                   RNZumoKit
