+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "react-native-zumo-kit",
-  "version": "4.0.0-beta.10",
+  "version": "4.1.0-alpha.2",
   "description": "ZumoKit is a Wallet as a Service SDK",
   "keywords": [
     "zumo",
@@ -31,7 +31,7 @@
     "react-native": "^0.62.14"
   },
   "dependencies": {
-    "zumokit": "4.0.0-beta.10",
+    "zumokit": "4.1.0-alpha.2",
     "decimal.js": "^10.2.0"
   },
   "devDependencies": {
