+++ src/Wallet.ts
@@ -1,14 +1,7 @@
-import { NativeModules } from 'react-native';
-import Decimal from 'decimal.js';
-import {
-  Transaction,
-  ComposedTransaction,
-  Exchange,
-  ComposedExchange,
-  ExchangeRate,
-  ExchangeSetting,
-} from 'zumokit/src/models';
-import { tryCatchProxy } from './utility/errorProxy';
+import { NativeModules } from "react-native";
+import Decimal from "decimal.js";
+import { ComposedTransaction, ComposedExchange } from "zumokit/src/models";
+import { tryCatchProxy } from "./utility/errorProxy";
 
 const {
   /** @internal */
@@ -95,111 +88,5 @@ export class Wallet {
     );
 
     return new ComposedTransaction(json);
-  }
-
-  /**
-   * Compose fiat transaction between users in Zumo ecosystem asynchronously.
-   * Refer to <a href="https://developers.zumo.money/docs/guides/send-transactions#internal-fiat-transaction">Send Transactions</a>
-   * guide for usage details.
-   *
-   * @param fromAccountId {@link  Account Account} identifier
-   * @param toAccountId   {@link  Account Account} identifier
-   * @param amount          amount in source account currency
-   * @param sendMax        send maximum possible funds to destination (defaults to false)
-   */
-  async composeInternalFiatTransaction(
-    fromAccountId: string,
-    toAccountId: string,
-    amount: Decimal | null,
-    sendMax = false
-  ) {
-    const json = await RNZumoKit.composeInternalFiatTransaction(
-      fromAccountId,
-      toAccountId,
-      amount ? amount.toString() : null,
-      sendMax
-    );
-
-    return new ComposedTransaction(json);
-  }
-
-  /**
-   * Compose transaction to nominated account asynchronously.
-   * Refer to <a href="https://developers.zumo.money/docs/guides/send-transactions#external-fiat-transaction">Send Transactions</a>
-   * guide for usage details.
-   *
-   * @param fromAccountId {@link  Account Account} identifier
-   * @param amount          amount in source account currency
-   * @param sendMax        send maximum possible funds to destination (defaults to false)
-   */
-  async composeTransactionToNominatedAccount(
-    fromAccountId: string,
-    amount: Decimal | null,
-    sendMax = false
-  ) {
-    const json = await RNZumoKit.composeTransactionToNominatedAccount(
-      fromAccountId,
-      amount ? amount.toString() : null,
-      sendMax
-    );
-
-    return new ComposedTransaction(json);
-  }
-
-  /**
-   * Submit a transaction asynchronously.
-   * Refer to <a href="https://developers.zumo.money/docs/guides/send-transactions#submit-transaction">Send Transactions</a>
-   * guide for usage details.
-   *
-   * @param composedTransaction Composed transaction retrieved as a result
-   *                            of one of the compose transaction methods
-   * @param metadata            Optional metadata that will be attached to transaction
-   */
-  async submitTransaction(composedTransaction: ComposedTransaction, metadata: any = null) {
-    const json = await RNZumoKit.submitTransaction(
-      composedTransaction.json,
-      JSON.stringify(metadata)
-    );
-
-    return new Transaction(json);
-  }
-
-  /**
-   * Compose exchange asynchronously.
-   * Refer to <a href="https://developers.zumo.money/docs/guides/make-exchanges#compose-exchange">Make Exchanges</a>
-   * guide for usage details.
-   *
-   * @param fromAccountId       {@link  Account Account} identifier
-   * @param toAccountId         {@link  Account Account} identifier
-   * @param amount              amount in deposit account currency
-   * @param sendMax             exchange maximum possible funds (defaults to false)
-   */
-  async composeExchange(
-    fromAccountId: string,
-    toAccountId: string,
-    amount: Decimal | null,
-    sendMax = false
-  ) {
-    const json = await RNZumoKit.composeExchange(
-      fromAccountId,
-      toAccountId,
-      amount ? amount.toString() : null,
-      sendMax
-    );
-
-    return new ComposedExchange(json);
-  }
-
-  /**
-   * Submit an exchange asynchronously.
-   * Refer to <a href="https://developers.zumo.money/docs/guides/make-exchanges#submit-exchange">Make Exchanges</a>
-   * guide for usage details.
-   *
-   * @param composedExchange Composed exchange retrieved as the result
-   *                          of {@link composeExchange} method
-   */
-  async submitExchange(composedExchange: ComposedExchange) {
-    const json = await RNZumoKit.submitExchange(composedExchange.json);
-    return new Exchange(json);
   }
 }
